fix: rename jpa test to relational test
