+++ backend/src/test/java/ca/bc/gov/backendstartapi/repository/seedlot/SeedlotOrchardRelationalTest.java
@@ -1,0 +1,47 @@
+package ca.bc.gov.backendstartapi.repository.seedlot;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import ca.bc.gov.backendstartapi.entity.embeddable.AuditInformation;
+import ca.bc.gov.backendstartapi.entity.seedlot.SeedlotOrchard;
+import ca.bc.gov.backendstartapi.entity.seedlot.idclass.SeedlotOrchardId;
+import ca.bc.gov.backendstartapi.enums.SeedlotStatusEnum;
+import ca.bc.gov.backendstartapi.repository.GeneticClassRepository;
+import ca.bc.gov.backendstartapi.repository.GeneticWorthRepository;
+import ca.bc.gov.backendstartapi.repository.SeedlotOrchardRepository;
+import ca.bc.gov.backendstartapi.repository.SeedlotRepository;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.transaction.annotation.Transactional;
+
+@DataJpaTest
+@Transactional
+class SeedlotOrchardRelationalTest extends SeedlotEntityRelationalTest {
+
+  private final SeedlotOrchardRepository repository;
+
+  @Autowired
+  protected SeedlotOrchardRelationalTest(
+      SeedlotRepository seedlotRepository,
+      GeneticClassRepository geneticClassRepository,
+      SeedlotOrchardRepository seedlotOrchardRepository,
+      GeneticWorthRepository geneticWorthRepository) {
+    super(seedlotRepository, geneticClassRepository, geneticWorthRepository);
+    repository = seedlotOrchardRepository;
+  }
+
+  @Test
+  void create() {
+    var seedlot = createSeedlot("00000", SeedlotStatusEnum.SUB);
+    var seedlotOrchard = new SeedlotOrchard(seedlot, "ABC");
+    seedlotOrchard.setPrimary(true);
+    seedlotOrchard.setAuditInformation(new AuditInformation("user1"));
+
+    repository.saveAndFlush(seedlotOrchard);
+
+    var savedSeedlotOrchard =
+        repository.findById(new SeedlotOrchardId(seedlot.getId(), seedlotOrchard.getOrchard()));
+    assertTrue(savedSeedlotOrchard.isPresent());
+  }
+}
