+++ backend/src/test/java/ca/bc/gov/backendstartapi/repository/seedlot/SmpMixGeneticQualityRelationalTest.java
@@ -3,12 +3,15 @@ package ca.bc.gov.backendstartapi.repository.seedlot;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import ca.bc.gov.backendstartapi.entity.SmpMix;
+import ca.bc.gov.backendstartapi.entity.SmpMixGeneticQuality;
 import ca.bc.gov.backendstartapi.entity.embeddable.AuditInformation;
+import ca.bc.gov.backendstartapi.entity.idclass.SmpMixGeneticQualityId;
 import ca.bc.gov.backendstartapi.entity.idclass.SmpMixId;
 import ca.bc.gov.backendstartapi.enums.SeedlotStatusEnum;
 import ca.bc.gov.backendstartapi.repository.GeneticClassRepository;
 import ca.bc.gov.backendstartapi.repository.GeneticWorthRepository;
 import ca.bc.gov.backendstartapi.repository.SeedlotRepository;
+import ca.bc.gov.backendstartapi.repository.SmpMixGeneticQualityRepository;
 import ca.bc.gov.backendstartapi.repository.SmpMixRepository;
 import java.math.BigDecimal;
 import org.junit.jupiter.api.Test;
@@ -18,18 +21,22 @@ import org.springframework.transaction.annotation.Transactional;
 
 @DataJpaTest
 @Transactional
-class SmpMixJpaTest extends SeedlotEntityJpaTest {
+class SmpMixGeneticQualityRelationalTest extends SeedlotEntityRelationalTest {
 
-  private final SmpMixRepository repository;
+  private final SmpMixRepository smpMixRepository;
+
+  private final SmpMixGeneticQualityRepository repository;
 
   @Autowired
-  SmpMixJpaTest(
+  SmpMixGeneticQualityRelationalTest(
       SeedlotRepository seedlotRepository,
       GeneticClassRepository geneticClassRepository,
       SmpMixRepository smpMixRepository,
+      SmpMixGeneticQualityRepository smpMixGeneticQualityRepository,
       GeneticWorthRepository geneticWorthRepository) {
     super(seedlotRepository, geneticClassRepository, geneticWorthRepository);
-    repository = smpMixRepository;
+    this.smpMixRepository = smpMixRepository;
+    repository = smpMixGeneticQualityRepository;
   }
 
   @Test
@@ -38,9 +45,21 @@ class SmpMixJpaTest extends SeedlotEntityJpaTest {
     var smpMix = new SmpMix(seedlot, 1, 1, null, new AuditInformation("user1"), 0);
     smpMix.setProportion(new BigDecimal(10));
 
-    repository.saveAndFlush(smpMix);
+    var savedSmpMix = smpMixRepository.save(smpMix);
+    var geneticWorth = geneticWorthRepository.findAll().get(0);
 
-    var savedSmpMix = repository.findById(new SmpMixId(seedlot.getId(), 1));
-    assertTrue(savedSmpMix.isPresent());
+    var smpMixGeneticQuality =
+        new SmpMixGeneticQuality(
+            smpMix, "GC", geneticWorth, new BigDecimal(10), true, new AuditInformation("user1"), 0);
+
+    repository.saveAndFlush(smpMixGeneticQuality);
+
+    var savedSmpMixGeneticQuality =
+        repository.findById(
+            new SmpMixGeneticQualityId(
+                new SmpMixId(seedlot.getId(), savedSmpMix.getParentTreeId()),
+                smpMixGeneticQuality.getGeneticTypeCode(),
+                smpMixGeneticQuality.getGeneticWorth().getGeneticWorthCode()));
+    assertTrue(savedSmpMixGeneticQuality.isPresent());
   }
 }
